Raspberry/requirements.txt
absl-py==2.0.0
av==11.0.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorzero==2.0
contourpy==1.2.0
cycler==0.12.1
easyocr==1.7.1
evdev==1.6.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gpiozero==2.0
grpcio==1.60.0
idna==3.6
imageio==2.33.1
imutils==0.5.4
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libusb1==3.1.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mfrc522==0.0.7
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.2
pandas==2.1.4
picamera==1.13
picamera2==0.3.16
pidng==4.0.9
piexif==1.1.3
Pillow==10.1.0
protobuf==4.21.2
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pyparsing==3.1.1
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.8.2
python-prctl==1.8.1
pytz==2023.3.post1
pyusb==1.2.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2
simplejpeg==1.7.2
six==1.16.0
spidev==3.6
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.232
urllib3==2.1.0
usb==0.0.83.dev0
v4l2-python3==0.3.2
Werkzeug==3.0.1
